% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{can.be.numeric}
\alias{can.be.numeric}
\title{Check whether a vector (usually a factor) can be cleanly converted to a numeric}
\usage{
can.be.numeric(x)
}
\arguments{
\item{x}{a vector}
}
\description{
From https://stackoverflow.com/a/47677916
}
