% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{codegen_load_all_sas_data}
\alias{codegen_load_all_sas_data}
\title{Create code to load all SAS datasets in a folder.}
\usage{
codegen_load_all_sas_data(dir, format = NULL)
}
\arguments{
\item{dir}{path to dataset folder}

\item{format}{path to format file}
}
\value{
NULL. Relevant code is printed to the console.
}
\description{
This is useful if you work with lots of separate SAS datasets spread out in the same folder.
}
\examples{
plot(c(1,2,3))
}
