% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{combine_two_elements_of_list}
\alias{combine_two_elements_of_list}
\title{Combines to summary statistics into one}
\usage{
combine_two_elements_of_list(lst, elem1, elem2, format_summary_stats)
}
\arguments{
\item{lst}{A DescrVar object}

\item{elem1, elem2}{Names of  summary statistics.}

\item{format_summary_stats}{Formatting options.}
}
\value{

}
\description{
Combines to summary statistics into one
}
\examples{

}
