% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{create_character_subtable.cont_summary}
\alias{create_character_subtable.cont_summary}
\title{Create subtables for continuous variables which will comprise the output table}
\usage{
\method{create_character_subtable}{cont_summary}(
  DescrVarObj,
  var_name,
  format_options,
  format_summary_stats,
  format_p
)
}
\arguments{
\item{var_name}{}
}
\value{

}
\description{
Create subtables for continuous variables which will comprise the output table
}
\examples{

}
