% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{create_character_subtable}
\alias{create_character_subtable}
\title{S3 dispatcher for subtable creation}
\usage{
create_character_subtable(DescrVarObj, ...)
}
\arguments{
\item{var_name}{}
}
\value{

}
\description{
S3 dispatcher for subtable creation
}
\examples{

}
