% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{create_printObj}
\alias{create_printObj}
\title{Create a DescrPrintObj which can be printed in the appropriate format by the appropriate method}
\usage{
create_printObj(DescrListObj, printFormat)
}
\arguments{
\item{DescrListObj}{A list obj.}

\item{printFormat}{What format?}
}
\value{

}
\description{
Create a DescrPrintObj which can be printed in the appropriate format by the appropriate method
}
\examples{

}
