% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/des.print.R
\name{des.print}
\alias{des.print}
\title{Create or print an output of a descriptive statistics table}
\usage{
des.print(dat, group, create = "pdf", file, index = T, fsize = 11, paired = F,
         nonparametric = F, var.equal = T, correct.cat = F, correct.wilcox = T,
         t.log = c(), which.col = c("groups", "total", "p-values"),
         group.min.size = F, n.or.miss = c("n", "miss", "miss.cat"), group.miss = F,
         percent.vertical = T, var.names, data.names = T, caption, tab.caption,
         landscape = F, pos.pagebr = NULL, label = NULL, digits.m = 1,
         digits.sd = 2, digits.qu = c(), digits.minmax = 1, digits.p = c(1,2),
         silent = T, ...)
}
\arguments{
\item{dat}{Data frame. The data set to be analyzed. Can contain continuous or factor (also ordered) variables.}

\item{group}{The (optional) grouping variable.\cr
Three options to specify it:\cr
1. The index of the group variable in \code{dat}.\cr
2. The variable name. It must be a variable in \code{dat}.\cr
3. As a vector with the same length as the number of rows in \code{dat}.\cr
4. Nothing. Then you need \code{which.col = ("total")}.\cr
The specified variable has to be a factor variable with two or more levels.
If not specified, a random grouping variable with 2 groups is used.}

\item{create}{Which output document should be produced (one of "pdf", "tex", "knitr","word" or "R"). Choose "custom" if you add more arguments see \code{...}.}

\item{file}{File name, which can included the directory (has to have the proper file extension, i.e. .pdf, .tex, or .docx). directory.
Only for \code{create == "R"} or \code{"knitr"} isn't a file necessary.}

\item{index}{Logical. Should the tests used be labeled by footnotes?}

\item{fsize}{The fontsize in the Output-Document.\cr
Default fontsize: For \code{create = "word"} or \code{"R"} fsize = 11 and for \code{create = "pdf"} fsize = 12}

\item{paired}{Logical. Should paired tests be applied? The groups must have the same length.}

\item{nonparametric}{Logical or vector of indices. If logical / vector of indices then all / only these continuous variables will be tested using non-parametric methods.}

\item{var.equal}{Logical. Should variances be assumed to be equal when applying t-tests?}

\item{correct.cat}{Logical. Should correction be used in chi-sqared tests (see \code{\link{chisq.test}})}

\item{correct.wilcox}{Logical. Should correction be used in wilcoxon tests (see \code{\link{wilcox.test}})}

\item{t.log}{Vector of indices: The variables for which the log of the original data
should be used when testing for a difference between the groups.}

\item{which.col}{Which columns should be provided ("groups", "total", "p-values")? Combinations are allowed. "groups" or "total" must be listed. Only "total" and "p-values" is not possible.
Type of p-value calculation see \code{\link{p.cat}} (categorical) or \code{\link{p.cont}} (continuous).}

\item{group.min.size}{For each variable, a p-value is only calculated if each non-empty group contains at least \code{group.min.size} observations for that variable.}

\item{group.non.empty}{For each variable, a p-value is only calculated if each group contains at least one observation for that variable.}

\item{cat.non.empty}{For categorical variables a p-value is only calculated if each category is non-empty.}

\item{n.or.miss}{Should the number of observations, missings for continuous variables, and/or missings for categorical variables be provided ("n", "miss", "miss.cat")? Combinations are allowed.}

\item{adaptive.miss}{Should the missing row be automatically omitted if there are not missings?}

\item{group.miss}{Logical. Should add a column for the Missings in group? For both cases in \code{total} there are only the observations without a missing in group.}

\item{percent.vertical}{Logical. Should "vertical" percentages for categorical variables be provided?}

\item{var.names}{Optional. Vector of names to be used in the table for the analyzed variables.}

\item{data.names}{Logical. If \code{var.names} are specified, should names as saved in \code{dat} be added in brackets?}

\item{caption}{Optional. Vector of names of the categories of group.\cr
If not specified, \code{levels(group)} will be used.}

\item{tab.caption}{Optional. The caption of the table.\cr
If not specified, "Descriptive statistics" will be written.}

\item{landscape}{Logical. Should the table be in landscape? Only useful for \code{create = "pdf"}.}

\item{pos.pagebr}{Vector of positions of the pagebreak in tex (or pdf). This is a bit fuzzy. It is the number of lines after a pagebreak should be done.\cr
If it is not specified, 45 will be used for "\code{landscape = FALSE}" and 30 will be used for "\code{landscape = TRUE}".}

\item{label}{Character. If tex (or knitr) in \code{create} is specified, the label of the table.}

\item{digits.m}{Number of digits for presentation in the table: For mean.}

\item{digits.sd}{Number of digits for presentation in the table: For standard deviation.}

\item{digits.qu}{Vector of numbers of digits for presentation in the table: For quantiles (if no value is specified it will be tried to provide a reasonable presentation).}

\item{digits.minmax}{Number of digits for presentation in the table: For minimum and maximum.}

\item{digits.p}{Vector with numbers of digits for presentation in the table: For percentages.
First vector element is number of digits for the first variable, second element for second variable and so on.}

\item{silent}{Logical. Should intermediate stages be shown (more for technical reasons)?}

\item{q.type}{an integer between 1 and 9 selecting one of the nine quantile algorithms detailed in \code{?quantile} to be used.}

\item{default.unordered.unpaired.test}{Any of c("Chisq", "Fisher_exact", "Fisher_boschloo"). Chooses the default test for categorical, unordered, unpaired variables.
 @param alignment
Specifies where the table shoould be aligned in the ouput document. Default is "c", other options are "l" and "r"}

\item{...}{further arguments to be passed to or from methods.}
}
\value{
Only by using \code{create == "R"} an R object will be returned. However, depending on the value of the create parameter either pdf, word, tex or an file optimized for use in connection with knitr will be created containing the descriptive statistics table and saved in a file as specified in the file parameter.
}
\description{
This is the main function for the user.
It can be used to generate a table of descriptive statistics with p-values obtained in tests for
difference between the groups. There are five options for the output: pdf, tex, knitr, word file
or an Output in R. knitr is not really useful as an output file but for the use within a knitr file.
}
\details{
The aim of this function is to help the user to create well-formated descriptive statistics tables.
The format can then be fine-tuned in the word, tex or knitr file itself. Only the pdf file is (more or less) unchangeable.
Even though it is supposed to be a good starting point for a descriptive analysis, care has to be taken when using the results and a detailed check of the results  might be necessary. For instance, in case of missing values in the group variable and \code{group.miss = FALSE} the respective observations will be omitted also in the total column. A warning will be displayed.
If no group variable is specified only the total column (see parameter \code{which.col}) will be returned.
}
\examples{
\dontrun{
infert
attach(infert)
#is.factor(education)
#is.factor(age)
#is.factor(parity)
#is.factor(induced)
#is.factor(case)
#infert$case<-as.factor(case)
#case<-as.factor(case)

#is.factor(spontaneous)
#is.factor(stratum)
#is.factor(pooled.stratum)

#we use case as Grouping variable
file <- "C:/Users/lunteren/Desktop/infert/DescriptiveStatisticTable.docx"

#Version 1
des.print(dat = infert, group = 5, create = "word", file=file, fsize = 10, var.names = c("education", "age", "parity","induced", "spontaneous", "stratum", "pooled.stratum"), caption = c("Group 1", "Group 2"))
#Version 2
des.print(dat = infert, group = case, create = "word", file=file, fsize = 10, var.names = c("education", "age", "parity","induced", "spontaneous", "stratum", "pooled.stratum"), caption = c("Group 1", "Group 2"))
#Version 3
group <- case
dat <- infert[,-5]
des.print(dat = dat, group = group, create = "word", file=file, fsize = 10, var.names = c("education", "age", "parity","induced", "spontaneous", "stratum", "pooled.stratum"), caption = c("Group 1", "Group 2"))

##Dataset with more then two groups
ChickWeight
attach(ChickWeight)

#is.factor(weight)
#is.factor(Time)
#is.factor(Chick)
#is.factor(Diet)

#we use Diet as Grouping variable

file <- "C:/Users/lunteren/Desktop/ChickWeight/DescriptiveStatisticTable.docx"

#Version 1
des.print(dat = ChickWeight, group = 4, create = "word", file = file, fsize = 10, var.names = c("weight", "Time", "Chick"), caption = c("Group 1", "Group 2", "Group 3", "Group 4"))
#Version 2
des.print(dat = ChickWeight, group = Diet, create = "word", file = file, fsize = 10, var.names = c("weight", "Time", "Chick"), caption = c("Group 1", "Group 2", "Group 3", "Group 4"))
#Version 3
group <- Diet
dat <- ChickWeight[,-4]
des.print(dat = dat, group = group, create = "word", file = file, fsize = 10, var.names = c("weight", "Time", "Chick"), caption = c("Group 1", "Group 2", "Group 3", "Group 4"))
}

}
\seealso{
\code{\link{descr}}\cr
\code{\link{f.r}}\cr
\code{\link{formatr}}\cr
\code{\link{inqur}}\cr
\code{\link{m.cat}}\cr
\code{\link{m.cont}}\cr
\code{\link{med.new}}\cr
\code{\link{minmax}}\cr
\code{\link{p.cat}}\cr
\code{\link{p.cont}}\cr
\link[flextable]{autofit}\cr
\link[flextable]{flextable}\cr
\link[flextable]{set_header_df}\cr
\link[flextable]{merge_h}\cr
\link[flextable]{merge_at}\cr
\link[flextable]{align}\cr
\link[flextable]{style}\cr
\link[flextable]{border}\cr
\link[flextable]{bold}\cr
\link[flextable]{width}\cr
\link[flextable]{height}\cr
\link[flextable]{body_add_flextable}\cr
\link[officer]{fp_text}\cr
\link[officer]{fp_cell}\cr
\link[officer]{fp_border}\cr
\link[officer]{read_docx}\cr
\link[xtable]{xtable}\cr
\link[tools]{texi2dvi}\cr
}
\author{
Lorenz Uhlmann, Csilla van Lunteren
}
