% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr}
\alias{descr}
\title{Create a descriptive statistics table}
\usage{
descr(
  dat,
  group = NULL,
  var_options = list(),
  group_labels = list(),
  summary_stats_cont = list(N = .N, Nmiss = .Nmiss, mean = .mean, sd = .sd, median =
    .median, Q1 = .Q1, Q3 = .Q3, min = .min, max = .max),
  summary_stats_cat = list(),
  format_p = scales::pvalue_format(),
  format_summary_stats = list(N = function(x) format(x, digits = 2, scientific = 3),
    Nmiss = function(x) format(x, digits = 2, scientific = 3), mean = function(x)
    format(x, digits = 2, scientific = 3), sd = function(x) format(x, digits = 2,
    scientific = 3), median = function(x) format(x, digits = 2, scientific = 3), Q =
    function(x) format(x, digits = 2, scientific = 3), minmax = function(x) format(x,
    digits = 2, scientific = 3), CI = function(x) format(x, digits = 2, scientific = 3)),
  format_options = list(print_p = T, print_CI = T, omit_Nmiss_if_0 = T,
    omit_missings_in_group = F, make_missing_a_category = F),
  test_options = list(paired = F, nonparametric = F, exact = F, indices = c(),
    include_group_missings_in_test = F, include_categorical_missings_in_test = F),
  ...
)
}
\arguments{
\item{dat}{Data frame or tibble. The data set to be analyzed. Can contain continuous or factor (also ordered) variables.}

\item{...}{Vector of the grouping variable.}
}
\value{
Depending on the value of the create parameter either pdf, word, tex, R
or an file optimized for use in connection with knitr will be created containing
the descriptive statistics table with the speak for the document to create in the following.
For example you choose \code{create="pdf"} then the table is written in \code{TeX}-Code.
Attention: the table has no caption and numbers of observations per group.
}
\description{
Generate a table of descriptive statistics with p-values obtained in tests
for difference between the groups.
}
\examples{
\dontrun{
}

}
\author{
Jan Meis, Lorenz Uhlmann, Csilla van Lunteren
}
