% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr_cont}
\alias{descr_cont}
\title{Create descriptive statistics for a continuous variable}
\usage{
descr_cont(
  var,
  group,
  var_name,
  summary_stats = c(),
  var_options = list(),
  test_options = list()
)
}
\arguments{
\item{group}{}
}
\value{

}
\description{
Create descriptive statistics for a continuous variable
}
\examples{

}
