% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescrTab.R
\docType{package}
\name{DescrTab}
\alias{DescrTab}
\title{DescrTab: A package to create descriptive statistics tables.}
\description{
The packages creates descriptive statistics such as mean, standard
deviation or relative and absolute frequencies grouped by a factor and also
provides p-values obtained from testing for differences between the groups.
}
\section{Want to learn?}{

Type in vignette(package="DescrTab")
}

