% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescrTab2.R
\docType{package}
\name{DescrTab2}
\alias{DescrTab2}
\title{DescrTab2: A package to create descriptive statistics tables.}
\description{
The packages creates descriptive statistics such as mean, standard
deviation or relative and absolute frequencies grouped by a factor and also
provides p-values obtained from testing for differences between the groups.
}
\section{Want to learn?}{

Type in vignette(package="DescrTab2")
}

