% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.factormean}
\alias{.factormean}
\title{Calculate mean of nonmissing values after coercing factor to numeric}
\usage{
.factormean(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate mean of nonmissing values after coercing factor to numeric
}
\examples{

}
