% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.max}
\alias{.max}
\title{Calculate maximum of nonmissing values}
\usage{
.max(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate maximum of nonmissing values
}
\examples{

}
