% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.mean}
\alias{.mean}
\title{Calculate mean of nonmissing values.}
\usage{
.mean(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate mean of nonmissing values.
}
\examples{

}
