% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.median}
\alias{.median}
\title{Calculate median of nonmissing values.}
\usage{
.median(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate median of nonmissing values.
}
\examples{

}
