% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.min}
\alias{.min}
\title{Calculate minimum of nonmissing values}
\usage{
.min(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate minimum of nonmissing values
}
\examples{

}
