% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.N}
\alias{.N}
\title{Calculate number of nonmissing values.}
\usage{
.N(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate number of nonmissing values.
}
\examples{

}
