% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.Nmiss}
\alias{.Nmiss}
\title{Calculate number of missing values.}
\usage{
.Nmiss(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate number of missing values.
}
\examples{

}
