% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.Q3}
\alias{.Q3}
\title{Calculate third quantile of nonmissing values.}
\usage{
.Q3(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate third quantile of nonmissing values.
}
\examples{

}
