% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{.sd}
\alias{.sd}
\title{Calculate standard deviation of nonmissing values.}
\usage{
.sd(var)
}
\arguments{
\item{var}{A variable (a vector).}
}
\value{

}
\description{
Calculate standard deviation of nonmissing values.
}
\examples{

}
