% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{escape_latex_symbols}
\alias{escape_latex_symbols}
\title{Escape LaTeX Symbols}
\usage{
escape_latex_symbols(tibl, doubleEscape = FALSE)
}
\arguments{
\item{tibl}{A \code{tibble} filled with characters}

\item{doubleEscape}{(logical) chooses between "\\" and "\\\\"#}
}
\value{
a \code{tibble} with appropriately escape LaTeX code
}
\description{
Escape LaTeX Symbols
}
