% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{extract_labels}
\alias{extract_labels}
\title{Extract the label attributes from data}
\usage{
extract_labels(dat)
}
\arguments{
\item{dat}{dataset or vector}
}
\value{
list of labels
}
\description{
Extract the label attributes from data
}
\examples{
a <- c(1,2)
attr(a, "label") <- "b"
unlabel(a)
}
