% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.r.R
\name{f.r}
\alias{f.r}
\title{Format numbers}
\usage{
f.r(x, k)
}
\arguments{
\item{x}{Number, which should be rounded.}

\item{k}{Number of decimal places.}
}
\value{
The rounded number is returned.
}
\description{
Rounds a number x to k decimal places.
}
\examples{
\dontrun
set.seed(12345)
x<-rnorm(1)
k<-3
f.r(x,k)
}

