% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{format_freqs}
\alias{format_freqs}
\title{Make pretty frequencies}
\usage{
format_freqs(
  numerator,
  denominator = 1,
  absolute_relative_frequency_mode = c("both", "only_absolute", "only_relative"),
  percent_accuracy = NULL,
  percent_suffix = "\%"
)
}
\arguments{
\item{numerator}{numeric}

\item{denominator}{numeric}

\item{absolute_relative_frequency_mode}{"both","only_absolute" or "only_relative"}

\item{percent_accuracy}{NULL or numeric}

\item{percent_suffix}{"\%" or "" (or something else)}
}
\value{
Character element of formatted frequencies
}
\description{
Make pretty frequencies
}
