% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.r.R
\name{formatr}
\alias{formatr}
\title{Format numbers}
\usage{
formatr(x, k = 3, cl.z = F)
}
\arguments{
\item{x}{Number, which should be rounded.}

\item{k}{Number of decimal places.}

\item{cl.z}{Logical. Should a uniform representation be made for all numbers that are not in the display area?}
}
\value{
The rounded number is returned, maybe with uniform representation.
}
\description{
Rounds a number x to k decimal places with a uniform representation (<0.001 for k=3) for all numbers that are not in the disabling area.
}
\examples{
\dontrun{
x<-c(0.1,0.01,0.001)
formatr(x, k=2, cl.z=T)
}

}
\seealso{
\code{\link{f.r}}
}
\author{
Lorenz Uhlmann, Csilla van Lunteren
}
