% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{get_groupNames}
\alias{get_groupNames}
\title{Extract group names from cat_summary or cont_summary object}
\usage{
get_groupNames(DescrVarObj)
}
\arguments{
\item{DescrVarObj}{A DescrVarObj}
}
\value{

}
\description{
Extract group names from cat_summary or cont_summary object
}
\examples{

}
