% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{guess_ID_variable}
\alias{guess_ID_variable}
\title{Make an educated guess about the name of the ID variable from a dataset}
\usage{
guess_ID_variable(dat, suppressWarnings = FALSE)
}
\arguments{
\item{dat}{a dataset with names (\code{list}, \code{data.frame}, \code{tibble})}

\item{suppressWarnings}{(logical) suppress warning messages if you know what you are dooing}
}
\value{
if exactly one possible
}
\description{
Make an educated guess about the name of the ID variable from a dataset
}
\examples{
dat <- data.frame(ID = c(1,2,3,4,5),
                 other = c(1,2,3,4,5))
guess_ID_variable(dat)
}
