% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{inMinipage2}
\alias{inMinipage2}
\title{Wrap cell text in minipage LaTeX environment}
\usage{
inMinipage2(x, width)

inMinipage2(x, width)
}
\arguments{
\item{x}{text to be placed in minipage}

\item{width}{width adjustment
https://stackoverflow.com/a/50892682}
}
\description{
Wrap cell text in minipage LaTeX environment

Wrap cell text in minipage LaTeX environment
}
