% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{knit_print.DescrList}
\alias{knit_print.DescrList}
\title{S3 override for knit_print function for DescrList objects.}
\usage{
\method{knit_print}{DescrList}(x, print_format = options("print_format")[[1]], silent = FALSE, ...)
}
\arguments{
\item{x}{a}

\item{print_format}{b}

\item{silent}{c}

\item{...}{abc}
}
\value{
outputs formatted table depending on the environment (.RMD) which it is called from
}
\description{
S3 override for knit_print function for DescrList objects.
}
