% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{knit_print.DescrPrint}
\alias{knit_print.DescrPrint}
\title{S3 override for knit_print function for DescrPrint objects.}
\usage{
\method{knit_print}{DescrPrint}(x, print_format = print_format, silent = silent, ...)
}
\arguments{
\item{x}{a}

\item{print_format}{b}

\item{silent}{c}

\item{...}{abc}
}
\value{
outputs formatted table depending on the environment (.RMD) which it is called from
}
\description{
S3 override for knit_print function for DescrPrint objects.
}
