% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{lapply_descr}
\alias{lapply_descr}
\title{Convenience function to apply descr to a list of datasets and print the results}
\usage{
lapply_descr(list, ...)
}
\arguments{
\item{list}{a list of datasets (tibbles or data.frames)}
}
\value{
something printable.
}
\description{
Convenience function to apply descr to a list of datasets and print the results
}
\examples{
l <- list()
for (i in 1:2){
  l <- append(l, list(iris))
}
lapply_descr(l, group="Species")

}
