% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.cat.R
\name{m.cat}
\alias{m.cat}
\title{Decide which test to use for the p-value calculation by categorial variables.}
\usage{
m.cat(group, paired = F, is.ordered = F)
}
\arguments{
\item{group}{Vector of the grouping variable.}

\item{paired}{Logical. Is the categorial Variable paired?}

\item{is.ordered}{Logical. Is the categorial Variable ordered?}
}
\value{
The name of the test is returned.
}
\description{
It choose the test which should use for the p-value calculation for categorial variables. The Wilcoxon-Test, McNemar-Test and the Chi-Squared-Test are possible.
}
\examples{
\dontrun{
m.cat(paired=T, is.ordered=T)
}
}
\author{
Lorenz Uhlmann, Csilla van Lunteren
}
