% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m.cont.R
\name{m.cont}
\alias{m.cont}
\title{Decide which test to use for the p-value calculation by continous variables.}
\usage{
m.cont(group, paired = F, is.ordered = F, nonparametric = F, t.log = F)
}
\arguments{
\item{group}{Vector of the grouping variable.}

\item{paired}{Logical. Is the categorial Variable paired?}

\item{is.ordered}{Logical. Is the categorial Variable ordered?}

\item{nonparametric}{Logical or vector of indices. If logical / vector of indices then all / only these continuous variables will be tested using non-parametric methods.}

\item{t.log}{Vector of indices: The variables for which the log of the original data should be used when testing for a difference between the groups.}
}
\value{
The Name of the Test is returned.
}
\description{
It choose the test which should use for the p-value calculation for continous variables. The Wilcoxon-Test, t-Test, Firedman-Test, Anova Typ III, Kruskal-Test and a Anova are possible.
}
\examples{
\dontrun{
m.cont(group=rep(1:4,25))
}

}
\author{
Lorenz Uhlmann, Csilla van Lunteren
}
