% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescrTab.R
\docType{data}
\name{madeUpData}
\alias{madeUpData}
\title{Data that is made to look like clinical data.}
\format{A data frame with 200 rows and 5 variables:
\describe{
  \itme{group}{control or treatment}
  \item{gender}{male or female}
  \item{age}{number between 15-90}
  \item{blood_hg}{Number between 50 and 150}
  \item{QoL_score}{sad, normal or happy}
  ...
}}
\usage{
madeUpData
}
\description{
For teaching purposes
}
\keyword{datasets}
