% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/med.new.R
\name{med.new}
\alias{med.new}
\title{Median value for continous variables}
\usage{
med.new(x, k = c())
}
\arguments{
\item{x}{Vector of the continous variable.}

\item{k}{Number of decimal places.
If not specified, the number of decimal places is 1 or 2.}
}
\value{
The median is returned.
}
\description{
Compute the median of continous variables.
}
\author{
Lorenz Uhlmann, Csilla van Lunteren
}
\seealso{
\code{\link{formatr}}
}

