% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minmax.R
\name{minmax}
\alias{minmax}
\title{Minimum and maximum value of continous variables}
\usage{
minmax(x, k = 1)
}
\arguments{
\item{x}{Vector of the continous variable.}

\item{k}{Number of decimal places.
If not specified, the number of decimal places is 1 or 2.}
}
\value{
\code{"minimum" -- "maximum"} is returned.
}
\description{
Compute the minimum and maximum of continous variables.
}
\examples{
\dontrun{
set.seed(12345)
x<-rnorm(100)

minmax(x,k=2)
}

}
\seealso{
\code{\link{formatr}}
}
\author{
Lorenz Uhlmann, Csilla van Lunteren
}
