% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{n_int_digits}
\alias{n_int_digits}
\title{Digits before decimal -1}
\usage{
n_int_digits(x)

n_int_digits(x)
}
\arguments{
\item{x}{}
}
\value{



}
\description{
Digits before decimal -1

Digits before decimal -1
}
\details{
https://stackoverflow.com/questions/47190693/count-the-number-of-integer-digits

https://stackoverflow.com/questions/47190693/count-the-number-of-integer-digits
}
