% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.cat.R
\name{p.cat}
\alias{p.cat}
\title{p-value calculator for categorical variables}
\usage{
p.cat(
  x,
  group,
  paired = F,
  is.ordered = F,
  correct.cat = F,
  correct.wilcox = T,
  index = c(),
  create = "tex",
  default.unordered.unpaired.test = "Chisq"
)
}
\arguments{
\item{x}{Vector of the categorial variable.}

\item{group}{Vector of the grouping variable.}

\item{paired}{Logical. Is the categorial Variable paired?}

\item{is.ordered}{Logical. Is the categorial Variable ordered?}

\item{correct.cat}{Logical. Should correction be used in chi-sqared tests (see \code{\link{chisq.test}})}

\item{correct.wilcox}{Logical. Should correction be used in wilcoxon tests (see \code{\link{wilcox.test}})}

\item{index}{Optional. Label for the footnote.
The footnotes aren't produced in this function.}

\item{create}{Which output document should be produced in the following step
(one of "pdf", "tex", "knitr", or "word").
Only usefull if \code{index} is not \code{NULL}.}
}
\value{
The p-value with index which test is ussed is returned.
}
\description{
Calculate the p-value for categorial variables.
The decision which test to use is equal to \code{\link{m.cat}}.
The p-value is calculated using one of the three tests:
Wilcoxon-Test, McNemar-Test, Chi-Squared-Test.
}
\details{
Wilcoxon-Test: A Test for a comparison of 2 (in)dependent, ordered samples.
(see \code{\link{wilcox.test}}).
Kruskal_wallis-Test: A Test for a comparison of more than 2 (in)dependent,
ordered samples. (see \code{\link{kruskal.test}}).
McNemar Test: A Test for a comparison of 2 dependent,
not ordered samples. (see \code{\link{mcnemar.test}}).
Chi-Squared Test: A Test for a comparison of 2 or more than 2 independent,
not ordered samples. (see \code{\link[DescTools]{CochranQTest}}).
Cochran's Q Test: A test for a comparison of 2 or more than 2 dependent,
not ordered samples. (see \code{\link{chisq.test}}).
}
\examples{
\dontrun{
p.cat(x=rep(1:5,20), group=rep(1:4,25))
}

}
\author{
Lorenz Uhlmann, Csilla van Lunteren
}
