% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p.cont.R
\name{p.cont}
\alias{p.cont}
\title{p-value calculator for continous variables}
\usage{
p.cont(x, group, paired = F, is.ordered = F, nonparametric = F, t.log = F, var.equal = F, index = c(), create = "tex")
}
\arguments{
\item{x}{Vector of the continous variable.}

\item{group}{Vector of the grouping variable.}

\item{paired}{Logical. Is the categorial Variable paired?}

\item{is.ordered}{Logical. Is the categorial Variable ordered?}

\item{nonparametric}{Logical. Should the continuous variable tested by using non-parametric methods.}

\item{t.log}{Logical. Should be used the log of the original data.}

\item{var.equal}{Logical. Should variances be assumed to be equal when applying t-tests?}

\item{index}{Optional. Label for the footnote.
The footnotes aren't produced in this function.}

\item{create}{Which output document should be produced in the following step (one of "pdf", "tex", "knitr", or "word").}
}
\value{
The p-value with index which test is ussed is returned.
author
Lorenz Uhlmann, Csilla van Lunteren
}
\description{
Calculate the p-value for continous variables. The decision which test to use is equal to \code{\link{m.cont}}.
The p-value is calculated using one of the four tests: Wilcoxon-Test, t-Test, Kruskal-Test, Anova.
}
\details{
Wilcoxon Test: A nonparametric Test for a comparison of 2 dependent samples. (see \code{\link{wilcox.test}}).
Mann-Whitney-U Test: A nonparametric Test for a comparison of 2 independent samples. (see \code{\link{wilcox.test}}).
t-Test: A parametric Test for a comparison of 2 (in)dependent samples. (see \code{\link{t.test}}).
Friedman-Test: A nonparametric Test for a comparison of more than 2 dependent samples. (see \code{\link{friedman.test}}).
Anova Type III: A parametric Test for a comparison of more than 2 dependent samples. (see \code{\link[car]{Anova}} with \code{}).
Kruskal-Wallis-Test: A nonparametric Test for a comparison of more than 2 independent samples. (see \code{\link{kruskal.test}}).
Anova: A parametric Test for a comparison of more than 2 independent samples. (see \code{\link{aov}}).
}
\seealso{
\link[nlme]{lme}
\link[car]{Anova}
examples
\dontrun
p.cont(x=rnorm(100,0,1), group=rep(1:4,25))
}

