% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{parse_formats}
\alias{parse_formats}
\title{SAS code parser for definitions of data formats.}
\usage{
parse_formats(path_to_format_definition)
}
\arguments{
\item{path_to_format_definition}{}
}
\value{

}
\description{
SAS code parser for definitions of data formats.
}
