% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{print_html}
\alias{print_html}
\title{Print descriptive statistics table with character entries in html format}
\usage{
print_html(DescrPrintObj, silent = F)
}
\arguments{
\item{DescrPrintObj}{A \code{printObj} object returned from \code{\link{create_printObj}}.}

\item{silent}{if TRUE, suppressed stdout output}
}
\value{

}
\description{
Print descriptive statistics table with character entries in html format
}
\examples{

}
