% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{print_numeric}
\alias{print_numeric}
\title{Print descriptive statistics table with numeric entries to console}
\usage{
print_numeric(
  DescrPrintObj,
  silent = F,
  n = 1000,
  width = NULL,
  n_extra = NULL,
  print_red_NA = F
)
}
\arguments{
\item{DescrPrintObj}{A \code{printObj} object returned from \code{\link{create_printObj}}.}

\item{silent}{if TRUE, suppressed stdout output}

\item{n}{tibble printing option}

\item{width}{tibble printing option}

\item{n_extra}{tibble printing option}

\item{print_red_NA}{if TRUE, prints red text for NA values in console}
}
\value{

}
\description{
Print descriptive statistics table with numeric entries to console
}
\examples{

}
