% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{print_test_names}
\alias{print_test_names}
\title{Prints all possible tests names}
\usage{
print_test_names()
}
\value{
Returns the names of all possible test names you can specify.
}
\description{
Prints all possible tests names
}
\examples{
print_test_names()
}
