% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{print_word}
\alias{print_word}
\title{Print descriptive statistics table with character entries in word format}
\usage{
print_word(DescrPrintObj, silent = F)
}
\arguments{
\item{DescrPrintObj}{A \code{printObj} object returned from \code{\link{create_printObj}}.}

\item{silent}{if TRUE, suppressed stdout output}
}
\value{

}
\description{
Print descriptive statistics table with character entries in word format
}
\examples{

}
