% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{read_sas_formatted}
\alias{read_sas_formatted}
\title{Convencience function to load SAS datasets}
\usage{
read_sas_formatted(path_to_data = NULL, path_to_format = NULL)
}
\arguments{
\item{path_to_data}{path to .sas7bdat file}

\item{path_to_format}{path to .sas7bcat file}
}
\value{
tibble with data
}
\description{
Convencience function to load SAS datasets
}
