% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{sig_test}
\alias{sig_test}
\title{Calculates a statistical significance test}
\usage{
sig_test(
  var,
  group = NULL,
  test_options = list(),
  test = NULL,
  var_name = NULL
)
}
\arguments{
\item{var}{A variable (a vector).}

\item{group}{A variable containing the grouping information.}

\item{test_options}{Named list containing test options.}

\item{test}{Name of a statistical test.}

\item{var_name}{Name of variable to be tested (only used in warning messages).}
}
\value{
A list of test test results.
}
\description{
Calculates a statistical significance test
}
\examples{
cont_var <- c(1, 2, 3)
sig_test(cont_var)
}
