% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{sigfig}
\alias{sigfig}
\title{Format number to a specified number of digits, considering threshold for usage of scientific notation}
\usage{
sigfig(
  x,
  digits = 3,
  scientific_high_threshold = 6,
  scientific_low_threshold = -6,
  force_0_behind_0 = FALSE
)

sigfig(
  x,
  digits = 3,
  scientific_high_threshold = 6,
  scientific_low_threshold = -6,
  force_0_behind_0 = FALSE
)
}
\arguments{
\item{digits}{}
}
\value{



}
\description{
Format number to a specified number of digits, considering threshold for usage of scientific notation

Format number to a specified number of digits, considering threshold for usage of scientific notation
}
