% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyDigits.R
\name{sigfig_gen}
\alias{sigfig_gen}
\title{Generator function for nice formatting functions}
\usage{
sigfig_gen(
  digits = 3,
  scientific_high_threshold = 6,
  scientific_low_threshold = -6,
  force_0_behind_0 = FALSE
)
}
\arguments{
\item{digits}{a}

\item{scientific_high_threshold}{a}

\item{scientific_low_threshold}{a}

\item{force_0_behind_0}{a}
}
\description{
Generator function for nice formatting functions
}
