% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{test_cat}
\alias{test_cat}
\title{calculate a statistical test for a numerical variable.}
\usage{
test_cat(var, group = NULL, test_options = NULL, test = NULL)
}
\arguments{
\item{var}{A variable (a vector).}

\item{group}{A variable containing the grouping information.}

\item{test_options}{Named list containing test options.}

\item{test}{Name of a statistical test.}
}
\value{
A list of test test results.
}
\description{
calculate a statistical test for a numerical variable.
}
\examples{
cat_var <- factor(c("a", "b", "c"))
test_cat(cat_var)

}
