% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{unlabel}
\alias{unlabel}
\title{Remove the label attribute from data}
\usage{
unlabel(dat)
}
\arguments{
\item{dat}{dataset or vector}
}
\value{
unlabelled dataset or vector
}
\description{
Remove the label attribute from data
}
\examples{
a <- c(1,2)
attr(a, "label") <- "b"
unlabel(a)
}
