% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{write_in_tmpfile_for_cran}
\alias{write_in_tmpfile_for_cran}
\title{Function that returns true in CRAN submission}
\usage{
write_in_tmpfile_for_cran()

write_in_tmpfile_for_cran()
}
\value{
TRUE for CRAN submission, FALSE otherwise

TRUE for CRAN submission, FALSE otherwise
}
\description{
Function that returns true in CRAN submission

Function that returns true in CRAN submission
}
