% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{combine_two_elements_of_list}
\alias{combine_two_elements_of_list}
\title{Does what the function name says}
\usage{
combine_two_elements_of_list(lst, elem1, elem2, format_summary_stats)
}
\arguments{
\item{elem2}{}
}
\value{

}
\description{
Does what the function name says
}
\examples{

}
