% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{create_character_subtable.cat_summary}
\alias{create_character_subtable.cat_summary}
\title{Title}
\usage{
\method{create_character_subtable}{cat_summary}(
  DescrVarObj,
  var_name,
  format_options,
  format_summary_stats,
  format_p
)
}
\value{

}
\description{
Title
}
\examples{

}
