% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{create_character_subtable}
\alias{create_character_subtable}
\title{Function to create (a part of a) nicely formatted table}
\usage{
create_character_subtable(
  DescrVarObj,
  var_name,
  format_options,
  format_summary_stats,
  format_p,
  reshape_rows
)
}
\arguments{
\item{DescrVarObj}{Variable object to be formatted}

\item{var_name}{(character) Name of the variable}

\item{format_options}{named list of options for formatting}

\item{format_summary_stats}{named list of summary statistcs}

\item{format_p}{formatting function for p-values}

\item{reshape_rows}{named list of row reshaping functions}
}
\description{
Function to create (a part of a) nicely formatted table
}
