% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{create_character_subtable}
\alias{create_character_subtable}
\title{S3 dispatcher for subtable creation}
\usage{
create_character_subtable(
  DescrVarObj,
  var_name,
  format_options,
  format_summary_stats,
  format_p
)
}
\arguments{
\item{DescrVarObj}{An object}

\item{var_name}{a}

\item{format_options}{b}

\item{format_summary_stats}{c}

\item{format_p}{d}
}
\value{

}
\description{
S3 dispatcher for subtable creation
}
\examples{

}
