% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{create_character_subtable}
\alias{create_character_subtable}
\title{Function to create (a part of a) nicely formatted table}
\usage{
create_character_subtable(
  DescrVarObj,
  var_name,
  format_options,
  format_summary_stats,
  format_p,
  reshape_rows
)
}
\description{
Function to create (a part of a) nicely formatted table
}
