% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{create_numeric_subtable.cat_summary}
\alias{create_numeric_subtable.cat_summary}
\title{Create sub-tables for categorical variables which will comprise the output table.}
\usage{
\method{create_numeric_subtable}{cat_summary}(
  DescrVarObj,
  var_name,
  format_options,
  format_summary_stats,
  format_p
)
}
\arguments{
\item{DescrVarObj}{An object returned from \code{\link{descr_cat}}.}

\item{var_name}{The name of the variable to be analyzed.}

\item{format_options}{b}

\item{format_summary_stats}{c}

\item{format_p}{d}
}
\value{

}
\description{
Create sub-tables for categorical variables which will comprise the output table.
}
\examples{

}
