% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr}
\alias{descr}
\title{Calculate descriptive statistics}
\usage{
descr(
  dat,
  group = NULL,
  var_options = list(),
  group_labels = list(),
  summary_stats_cont = list(N = DescrTab2:::.N, Nmiss = DescrTab2:::.Nmiss, mean =
    DescrTab2:::.mean, sd = DescrTab2:::.sd, median = DescrTab2:::.median, Q1 =
    DescrTab2:::.Q1, Q3 = DescrTab2:::.Q3, min = DescrTab2:::.min, max =
    DescrTab2:::.max),
  summary_stats_cat = list(),
  format_p = scales::pvalue_format(),
  format_summary_stats = list(N = function(x) format(x, digits = 2, scientific = 3),
    Nmiss = function(x) format(x, digits = 2, scientific = 3), mean = function(x)
    format(x, digits = 2, scientific = 3), sd = function(x) format(x, digits = 2,
    scientific = 3), median = function(x) format(x, digits = 2, scientific = 3), Q =
    function(x) format(x, digits = 2, scientific = 3), minmax = function(x) format(x,
    digits = 2, scientific = 3), CI = function(x) format(x, digits = 2, scientific = 3)),
  format_options = list(print_p = T, print_CI = T, omit_Nmiss_if_0 = T,
    omit_missings_in_group = F, make_missing_a_category = F),
  test_options = list(paired = F, nonparametric = F, exact = F, indices = c(),
    include_group_missings_in_test = F, include_categorical_missings_in_test = F),
  ...
)
}
\arguments{
\item{dat}{Data frame or tibble. The data set to be analyzed. Can contain continuous or factor (also ordered) variables.}

\item{group}{name (as character) of the group variable in dat}

\item{var_options}{named list of calculation, testing and formatting options for each variable in dat}

\item{group_labels}{named list of labels for the levels of the group variable in dat}

\item{summary_stats_cont}{named list of summary stats to be used for numeric variables}

\item{summary_stats_cat}{named list of summary stats to be used for categorical variables}

\item{format_p}{formatting function for p-values}

\item{format_summary_stats}{named list of formatting functions for summary statistics}

\item{format_options}{named list of formatting options}

\item{test_options}{named list of test options}

\item{...}{further argument passed along}
}
\value{
Returns a A \code{DescrList} object, which is a named list of descriptive statistics
which can be passed along to the print function to create
pretty summary tables.
}
\description{
Generate a list of descriptive statistics with p-values obtained in tests
for difference between the groups.
}
\examples{
descr(iris)
DescrList <- descr(iris)
DescrList$variables$Sepal.Length$Total$mean
print(DescrList)

}
\author{
Jan Meis, Lorenz Uhlmann, Csilla van Lunteren
}
