% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr}
\alias{descr}
\title{Calculate descriptive statistics}
\usage{
descr(
  dat,
  group = NULL,
  group_labels = list(),
  var_labels = list(),
  var_options = list(),
  summary_stats_cont = list(N = DescrTab2:::.N, Nmiss = DescrTab2:::.Nmiss, mean =
    DescrTab2:::.mean, sd = DescrTab2:::.sd, median = DescrTab2:::.median, Q1 =
    DescrTab2:::.Q1, Q3 = DescrTab2:::.Q3, min = DescrTab2:::.min, max =
    DescrTab2:::.max),
  summary_stats_cat = list(),
  format_summary_stats = list(N = function(x) format(x, digits = 2, scientific = 3),
    Nmiss = function(x) format(x, digits = 2, scientific = 3), mean = function(x)
    format(x, digits = 2, scientific = 3), sd = function(x) format(x, digits = 2,
    scientific = 3), median = function(x) format(x, digits = 2, scientific = 3), Q1 =
    function(x) format(x, digits = 2, scientific = 3), Q3 = function(x) format(x, digits
    = 2, scientific = 3), min = function(x) format(x, digits = 2, scientific = 3), max =
    function(x) format(x, digits = 2,      scientific = 3), CI = function(x) format(x,
    digits = 2, scientific = 3)),
  format_p = scales::pvalue_format(),
  format_options = list(print_p = T, print_CI = T, combine_mean_sd = F, omit_Nmiss_if_0
    = T, omit_missings_in_group = T, categorical_missing_percent_mode =
    c("no_missing_percent", "missing_as_regular_category",
    "missing_as_separate_category")),
  test_options = list(paired = F, nonparametric = F, exact = F, indices = c(),
    include_group_missings_in_test = F, include_categorical_missings_in_test = F,
    test_override = NULL),
  reshape_rows = list(`Q1 - Q3` = list(args = c("Q1", "Q3"), fun = function(Q1, Q3)
    paste0(Q1, " -- ", Q3)), `min - max` = list(args = c("min", "max"), fun =
    function(min, max) paste0(min, " -- ", max))),
  ...
)
}
\arguments{
\item{dat}{Data frame or tibble. The data set to be analyzed. Can contain continuous or factor (also ordered) variables.}

\item{group}{name (as character) of the group variable in dat.}

\item{group_labels}{named list of labels for the levels of the group variable in dat.}

\item{var_labels}{named list of variable labels.}

\item{var_options}{A named list of lists. For each variable, you can have special options that apply only to that variable.
These options are specified in this argument. See the details and examples for more explanation.}

\item{summary_stats_cont}{named list of summary statistic functions to be used for numeric variables.}

\item{summary_stats_cat}{named list of summary statistic function to be used for categorical variables.}

\item{format_summary_stats}{named list of formatting functions for summary statistics.}

\item{format_p}{formatting function for p-values.}

\item{format_options}{named list of formatting options.}

\item{test_options}{A named list of test options.}

\item{reshape_rows}{A named list of lists. Describes how to combine different summary statistics into the same row.}

\item{...}{further argument to be passed along}
}
\value{
Returns a A \code{DescrList} object, which is a named list of descriptive statistics
which can be passed along to the print function to create
pretty summary tables.
}
\description{
Generate a list of descriptive statistics. By default, the function calculates summary statistics such as mean,
standard deviation, quantiles, minimum and maximum for continuous variables and relative and absolute frequencies
for categorical variables. Also calculates p-values for an appropriately chosen statistical test.
For two-group comparisons, confidence intervals for appropriate summary measures of group differences are calculated aswell.
}
\section{Labels}{

\code{group_labels} and \code{var_labels} need to be named lists of character elements. The names of the list elements have to match the variable
names in your dataset. The values of the list elements are the labels that will be assigned to these variables when printing.
}

\section{Custom summary statistics}{

\code{summary_stats_cont} and \code{summary_stats_cat} are both named lists of functions. The names of the list elements are
what will be displayed in the leftmost column of the descriptive table. These functions should take a vector and return
a value. \cr
Each summary statistic has to have an associated formatting function in the \code{format_summary_stats} list.
The functions in \code{format_summary_stats} take a numeric value and convert it to a character string, e.g. 0.2531235 -> "0.2". \cr
The \code{format_p} function converts p-values to character strings, e.g. 0.05 -> "0.05" or 0.000001 -> "<0.001".
}

\section{Formatting options}{

Further formatting options can be specified in the \code{format_options} list. It contains the following members:
\itemize{
\item{\code{print_p}}{ (logical) controls whether to print the p-value column.}
\item{\code{print_CI}}{ (logical) controls whether to print the confidence intervals for group-differences.}
\item{\code{combine_mean_sd}}{ (logical) controls whether to combine the mean and sd row into one mean ± sd row. This is a
shortcut argument for the specification of an appropriate entry in the \code{reshape_rows} argument.}
\item{\code{omit_Nmiss_if_0}}{ (logical)  controls whether to omit the Nmiss row in continuous variables there are no missings in the variable.}
\item{\code{omit_missings_in_group}}{ (logical)  controls whether to omit all observations where the group variable is missing.}
\item{\code{categorical_missing_percent_mode}}{ (character)  controls how to display percentages in categorical variables with a (Missing) category.
It may be set to one of the following options:
\itemize{
\item{\code{"no_missing_percent"}}{ omits a percentage in the missing category entirely.}
\item{\code{"missing_as_regular_category"}}{ treats (Missing) as a regular category for \%-calculation
This means that if You have three categories: "A" with 10 counts, "B" with 10 counts and "(Missing)" with 10 counts,
they will become "A": 10 (33\%), "B": 10 (33\%), "(Missing)": 10 (33\% purposes.)}

\item{\code{"missing_as_separat_category"}}{ calculates (Missing) percentages with respect to
all observations (i.e. #(Missing) / N), but calculates all other catetgory percentages with respect to the non-missing
observations (e.g. #A / N_nonmissing). This means that if You have three categories: "A" with 10 counts, "B" with 10 counts
and "(Missing)" with 10 counts, they will become "A": 10 (50\%), "B": 10 (50\%), "(Missing)": 10 (33\%)}
  }
 }
}
}

\section{Test options}{

\code{test_options} is a named list with test options. It's members \code{paired},  \code{nonparametric}, and
\code{exact} (logicals) control which test in the corresponding situation. For details, check out the vignette:
\code{vignette("test_choice_tree", "DescrTab2")}.
}

\section{Customization for single variables}{

The \code{var_options} list can be used to conduct customizations that should only apply to a single variable and leave
the rest of the table unchanged. \cr
\code{var_options} is a list of named lists. This means that each member of \code{var_options} is itself a list again.
The names of the list elements of \code{var_options} determine the variables to which the options will apply.
Let's say you have an \code{age} variable in your dataset. To change descr options only for \code{age}, you will need to pass
a list of the form \code{var_options = list(age = list(<Your options here>))}. \cr
You can replace \code{<Your options here>} with the following options:
\itemize{
\item{\code{label}}{ a character string containing the label for the variable}
\item{\code{summary_stats}}{ a list of summary statistics. See section "Custom summary statistics"}
\item{\code{format_summary_stats}}{ a list of formatting functions for summary statistics. See section "Custom summary statistics"}
\item{\code{format_p}}{ a function to format p-values. See section "Custom summary statistics"}
\item{\code{format_options}}{ a list of formatting options. See section "Formatting options"}
\item{\code{test_options}}{ a list of test options. See section "Test options"}
\item{\code{test_override}}{ manually specify the name of the test you want to apply. You can see a list of choices
by typing \code{print_test_names()}. Possible choices are:
\itemize{
\item{\code{"Cochrans Q test"}}{ }
\item{\code{"McNemars test"}}{ }
\item{\code{"Chi-squared goodness-of-fit test"}}{ }
\item{\code{"Pearsons chi-squared test"}}{ }
\item{\code{"Exact McNemars test"}}{ }
\item{\code{"Boschloos test"}}{ }
\item{\code{"Wilcoxon one-sample signed-rank test"}}{ }
\item{\code{"Mann-Whitney U test"}}{ }
\item{\code{"Kruskal-Wallis one-way ANOVA"}}{ }
\item{\code{"Students paired t-test"}}{ }
\item{\code{"Mixed model ANOVA"}}{ }
\item{\code{"Students one-sample t-test"}}{ }
\item{\code{"Welchs two-sample t-test"}}{ }
\item{\code{"F-test (ANOVA)"}}{ }
  }
 }
}
}

\section{Combining rows}{

The \code{reshape_rows} argument offers a framework for combining multiple rows of the output table into a single one.
\code{reshape_rows} is a named list of lists. The names of it's member-lists determine the name that will be displayed
as the name of the combined summary stats in the table (e.g. "mean ± sd "). The member lists need to contain two
elements: \code{args}, contains the names of the summary statistics to be combined as characters, and
\code{fun} which contains a function to combine these summary stats. The argument names of this function need to match
the character strings specified in \code{args}. Check out the default options for an exemplary definition.
}

\examples{
descr(iris)
DescrList <- descr(iris)
DescrList$variables$results$Sepal.Length$Total$mean
print(DescrList)
descr(iris, "Species")


}
