% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{descr_cont}
\alias{descr_cont}
\title{Create descriptive statistics for a continuous variable}
\usage{
descr_cont(
  var,
  group,
  var_name,
  summary_stats = c(),
  var_options = list(),
  test_options = list()
)
}
\arguments{
\item{var}{The variable to be analyzed (a vector).}

\item{group}{A vector of the same length as var containing the group assignment for the values in var.}

\item{var_name}{The name of the variable to be analyzed.}

\item{summary_stats}{named list of summary statistic functions.}

\item{var_options}{named list of calculation, testing and formatting options}

\item{test_options}{named list of test options.}
}
\value{

}
\description{
Create descriptive statistics for a continuous variable
}
\examples{

}
