% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescrTab2.R
\docType{package}
\name{DescrTab2}
\alias{DescrTab2}
\title{DescrTab2}
\description{
Publication quality descriptive statistics tables with R
}
\details{
Provides functions to create descriptive statistics tables for continuous and categorical variables.
By default, summary statistics such as mean, standard deviation, quantiles, minimum and maximum for continuous
variables and relative and absolute frequencies for categorical variables are calculated. 'DescrTab2' features a sophisticated algorithm to
choose appropriate test statistics for your data and provides p-values. On top of this, confidence intervals for group
differences of appropriated summary measures are automatically produces for two-group comparison.
Tables generated by 'DescrTab2' can be integrated in a variety of document formats, including .html, .tex and .docx documents.
'DescrTab2' also allows printing tables to console and saving table objects for later use.

Check out our documentation online: https://imbi-heidelberg.github.io/DescrTab2/
or browse the help files in the Rstudio viewer. You can access the vignettes by typing: \code{browseVignettes("DescrTab2")}

The most important function you probably want to check out is called \code{\link{descr}}.
}
