% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescrTab2.R
\name{.onLoad}
\alias{.onLoad}
\title{Load LaTeX packages}
\usage{
.onLoad(libname = find.package("kableExtra"), pkgname = "kableExtra")
}
\description{
Load LaTeX packages
}
\details{
Thanks to Hao Zhu and his package \link[kableExtra]{kableExtra}.
}
\author{
Hao Zhu
}
