% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{escape_latex_symbols}
\alias{escape_latex_symbols}
\title{Escape special characters in a  tibble containing LaTeX code}
\usage{
escape_latex_symbols(tibl)
}
\arguments{
\item{tibl}{A tibble containing LaTeX code.}
}
\value{

}
\description{
Escape special characters in a  tibble containing LaTeX code
}
\examples{

}
