% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyDigits.R
\name{format_freqs}
\alias{format_freqs}
\title{Formatting function for absolute and relative frequencies}
\usage{
format_freqs(
  numerator,
  denominator = 1,
  absolute_relative_frequency_mode = c("both", "only_absolute", "only_relative"),
  percent_accuracy = NULL,
  percent_suffix = "\%"
)
}
\arguments{
\item{numerator}{(numeric) numerator for \% calculations}

\item{denominator}{(numeric) denominator for \% calculations}

\item{absolute_relative_frequency_mode}{one of \code{c("both", "only_absolute", "only_relative")}.
"both" will print "Absolute Freq. (Relative Freq. \%)", the other options work accordingly.}

\item{percent_accuracy}{NULL or numeric. Refer to the \code{accuracy} argument in
\code{\link[scales]{percent}}.}

\item{percent_suffix}{usually "\%" or "". Refer to the \code{suffix} argument in
\code{\link[scales]{percent}}.}
}
\value{
string of formatted frequencies
}
\description{
Formatting function for absolute and relative frequencies
}
