% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latex.R
\name{in_minipage}
\alias{in_minipage}
\title{Wrap cell text in minipage LaTeX environment with stretchy space}
\usage{
in_minipage(text, width, numEscapes = 1, strechSpace = FALSE)
}
\arguments{
\item{text}{text to be placed in minipage}

\item{width}{width adjustment}

\item{numEscapes}{(logical) chooses between "\\" and "\\\\"#}

\item{strechSpace}{(logical) will add strethcy space}
}
\value{
appropriate LaTeX code
}
\description{
Wrap cell text in minipage LaTeX environment with stretchy space
}
\references{
\url{https://stackoverflow.com/a/50892682}
}
