% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inqur.R
\name{inqur}
\alias{inqur}
\title{Calculation of the first and third quantile}
\usage{
inqur(x, k = c())
}
\arguments{
\item{x}{Vector of continuous cariables.}

\item{k}{Optional. The number of decimal places.
If not specified, the number of decimal places is 1 or 2.}
}
\value{
The first and third quantil is returned.
}
\description{
It calculate the first an third quantil of continuous variables.
}
\examples{
\dontrun{
set.seed(12345)
x<-rnorm(100)

inqur(x, k = c(3))
}

}
\author{
Lorenz Uhlmann, Csilla van Lunteren
seealso
\code{\link{formatr}}
}
