% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{knit_print.DescrList}
\alias{knit_print.DescrList}
\title{S3 override for knit_print function for DescrList objects.}
\usage{
\method{knit_print}{DescrList}(x, print_format = options("print_format")[[1]], silent = FALSE, ...)

\method{knit_print}{DescrList}(x, print_format = options("print_format")[[1]], silent = FALSE, ...)
}
\value{



}
\description{
S3 override for knit_print function for DescrList objects.

S3 override for knit_print function for DescrList objects.
}
