% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{knit_print.DescrPrint}
\alias{knit_print.DescrPrint}
\title{S3 override for knit_print function for DescrPrint objects.}
\usage{
\method{knit_print}{DescrPrint}(x, print_format = print_format, silent = silent, ...)

\method{knit_print}{DescrPrint}(x, print_format = print_format, silent = silent, ...)
}
\arguments{
\item{x}{a}

\item{print_format}{b}

\item{silent}{c}

<<<<<<< HEAD
\item{...}{abc}
}
\value{
outputs formatted table depending on the environment (.RMD) which it is called from
=======


>>>>>>> 84e37d303bfcd68b0cacacb1d30b76e570f01328
}
\description{
S3 override for knit_print function for DescrPrint objects.

S3 override for knit_print function for DescrPrint objects.
}
