% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{knit_print.DescrPrint}
\alias{knit_print.DescrPrint}
\title{S3 override for knit_print function for DescrPrint objects.}
\usage{
\method{knit_print}{DescrPrint}(x, print_format = print_format, silent = silent, ...)

\method{knit_print}{DescrPrint}(x, print_format = print_format, silent = silent, ...)
}
\value{



}
\description{
S3 override for knit_print function for DescrPrint objects.

S3 override for knit_print function for DescrPrint objects.
}
