% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{list_freetext_markdown}
\alias{list_freetext_markdown}
\title{Create a markdown listing from a character dataset}
\usage{
list_freetext_markdown(dat)
}
\arguments{
\item{dat}{a character \code{data.frame} or \code{tibble}.}
}
\value{
string containing markdown code listing all nonempty free text in the dataset
}
\description{
Create a markdown listing from a character dataset
}
\examples{
dat  <- data.frame(Freetext = c("Some text", "More text"))
list_freetext_markdown(dat)
# use inside a .Rmd document like this:
# `r list_freetext_markdown(dat)`
}
