% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prettyDigits.R
\name{n_int_digits}
\alias{n_int_digits}
\title{Digits before decimal -1}
\usage{
n_int_digits(x)
}
\arguments{
\item{x}{a}
}
\value{
a
}
\description{
Digits before decimal -1
}
\details{
https://stackoverflow.com/questions/47190693/count-the-number-of-integer-digits
}
