% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{print.DescrList}
\alias{print.DescrList}
\title{S3 override for print function for DescrList objects}
\usage{
\method{print}{DescrList}(
  DescrListObj,
  printFormat = options("DescrTabFormat")[[1]],
  var_options = list(),
  group_labels = list(),
  format_p = NULL,
  format_summary_stats = list(),
  format_options = list(),
  ...
)
}
\arguments{
\item{printFormat}{Possible values: "console" (default), "tex", "html", "word", "numeric"}

\item{...}{}
}
\value{

}
\description{
S3 override for print function for DescrList objects
}
\examples{

}
