% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{print.DescrPrint}
\alias{print.DescrPrint}
\title{S3 override for print function for DescrPrint objects}
\usage{
\method{print}{DescrPrint}(x, print_format = options("print_format")[[1]], silent = FALSE, ...)
}
\arguments{
\item{x}{A \code{DescrList} object returned from \code{\link{descr}}.}

\item{print_format}{Possible values: "console" (default), "tex", "html", "word", "numeric"}

\item{silent}{I TRUE, suppresses output to stdout.}

\item{...}{further arguments to be passed along to print method}
}
\value{
A DescrPrint object which can be printed in various formats.
}
\description{
S3 override for print function for DescrPrint objects
}
\examples{
descr(iris)
DescrPrint <- print(descr(iris))
DescrPrint$variables$results$Sepal.Length$Total$mean
print(DescrPrint)

}
\keyword{internal}
