% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{print_numeric}
\alias{print_numeric}
\title{Title}
\usage{
print_numeric(
  DescrPrintObj,
  n = 1000,
  width = NULL,
  n_extra = NULL,
  print_red_NA = F
)
}
\arguments{
\item{print_red_NA}{}
}
\value{

}
\description{
Title
}
\examples{

}
