% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{read_sas_formatted}
\alias{read_sas_formatted}
\title{Convencience function to load sas datasets}
\usage{
read_sas_formatted(path_to_dat = NULL, path_to_format = NULL)
}
\arguments{
\item{path_to_dat}{path to .sas7bdat file}

\item{path_to_format}{path to .sas7bcat file}
}
\value{
tibble with data
}
\description{
Convencience function to load sas datasets
}
\examples{
path_to_dat <- system.file("examples", "testsas.sas7bdat", package = "DescrTab2")
pat_to_format <- system.file("examples", "formats.sas7bcat", package = "DescrTab2")
read_sas(path_to_dat, pat_to_format)
}
