% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{sigfig_gen}
\alias{sigfig_gen}
\title{Generator function for nice formatting functions}
\usage{
sigfig_gen(
  digits = 3,
  scientific_high_threshold = 6,
  scientific_low_threshold = -6,
  force_0_behind_0 = FALSE
)

sigfig_gen(
  digits = 3,
  scientific_high_threshold = 6,
  scientific_low_threshold = -6,
  force_0_behind_0 = FALSE
)
}
\arguments{
\item{digits}{}
}
\value{



}
\description{
Generator function for nice formatting functions

Generator function for nice formatting functions
}
\examples{
plot(c(1,2,3))
}
