% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{test_cat}
\alias{test_cat}
\title{Calculate a statistical test for a categorical variable.}
\usage{
test_cat(
  var,
  group = NULL,
  test_options = list(),
  test = NULL,
  var_name = NULL
)
}
\arguments{
\item{var}{A variable (a vector).}

\item{group}{A variable containing the grouping information.}

\item{test_options}{Named list containing test options.}

\item{test}{Name of a statistical test.}

\item{var_name}{Name of variable to be tested (only used in warning messages).}
}
\value{
A list of test test results.
}
\description{
Calculate a statistical test for a categorical variable.
}
\examples{
cat_var <- factor(c("a", "b", "c"))
test_cat(cat_var)

}
