% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{test_cont}
\alias{test_cont}
\title{calculate a statistical test for a numerical variable.}
\usage{
test_cont(var, group, test_options, test = NULL)
}
\arguments{
\item{var}{A variable (a vector).}

\item{group}{A variable containing the grouping information.}

\item{test_options}{Named list containing test options.}

\item{test}{Name of a statistical test.}
}
\value{

}
\description{
calculate a statistical test for a numerical variable.
}
\examples{

}
