% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descr.R
\name{test_cont}
\alias{test_cont}
\title{Calculate a statistical test for a numerical variable.}
\usage{
test_cont(var, group = NULL, test_options = NULL, test = NULL, var_name = NULL)
}
\arguments{
\item{var}{A variable (a vector).}

\item{group}{A variable containing the grouping information.}

\item{test_options}{Named list containing test options.}

\item{test}{Name of a statistical test.}

\item{var_name}{Name of variable to be tested (only used in warning messages).}
}
\value{
A list of test test results.
}
\description{
Calculate a statistical test for a numerical variable.
}
\examples{
cont_var <- c(1,2,3)
test_cont(cont_var)

}
