% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadData.R
\name{unlabel}
\alias{unlabel}
\title{Remove the label attribute from data}
\usage{
unlabel(dat)
}
\arguments{
\item{dat}{data in the form of a \code{\link[base]{list}}, \code{\link[base]{data.frame}}
or \code{\link[tibble]{tibble}}, or a vector}
}
\value{
data with the labels removed
}
\description{
Remove the label attribute from data
}
